requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pypdf2==3.0.1
sentence-transformers==3.2.1
scikit-learn==1.5.2
numpy==2.1.3
